refac(app): rename root_folder_selector component to workpsace_selector
