--- apps/app/src/components/general/root_folder_selector/states.svelte.ts
@@ -1,3 +1,0 @@
-export const root_folder_picker_dialog_state: { open: boolean } = $state({
-  open: false,
-});
