+++ apps/app/src/components/general/workspace_selector/states.svelte.ts
@@ -1,0 +1,3 @@
+export const workspace_picker_dialog_open_state = $state({
+  data: false,
+});
